pyproject.toml
[build-system]
requires = ["setuptools>=63", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synapse"
version = "0.1.0"
description = "A PyTorch Lightning wrapper for high energy physics"
readme = "README.md"
requires-python = "~=3.11"
authors = [{name = "Shudong Wang", email = "907265139@qq.com"}]
license = {text = "MIT"}

classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "awkward~=2.8.3",
    "numpy~=2.2.6",
    "PyYAML~=6.0.2",
    "uproot~=5.6.2",
    "colorlog~=6.9.0",
    "torch~=2.7.0",
    "lightning~=2.4.0",
    "tensorboard~=2.19.0",
    "scikit-learn~=1.7.0",
    "onnx~=1.18.0",
]

[project.entry-points.console_scripts]
synapse = "synapse.main:main"
synapse-convert-hhml = "synapse.widgets.convert_ntuple_hhml:main"

[project.urls]
"Homepage" = "https://github.com/Shudong-Wang/Synapse"
"Bug Tracker" = "https://github.com/Shudong-Wang/Synapse/issues"

[tool.setuptools.packages.find]
where = ["src"]
